test(regressions): split the NovelAI caption-cap guard into its own file
The Illustrator caption regression imported both the image-generation and
agent-executor module graphs, which pushed it against the runner's fixed
30-second budget on a loaded machine. Each file now loads one graph.
